# Cache `Publisher` objects to speed up Sphinx