I don't think this is a big enough change to justify being a "feature". It does seem to be quite a big speed-up though (at least 20% faster).

Tests with `builddoc.yml` show 55s on `5.x` and 35s on this branch (~36% faster).

Tests on the main CI (Python 3.10, ubuntu) show 211s on `5.x` and 160s on this branch (~24% faster).

A

### Feature or Bugfix
- Bugfix