@AA-Turner 2022-05-07
@tk0miya I merged the function into `Builder.read_doc`, I think that's what you meant?

> what is the merit of cloning?

I added a longer comment -- Docutils' publisher reuses the settings object, but in `Builder.write_doctree` we edit it for pickleability. I want to update Docutils to allow more of the behaviour Sphinx has to patch to happen natively, but that will take a little time.

A